+++ Source/FloatSliderLFOControl.h
@@ -120,9 +120,9 @@ protected:
 
 private:
    void UpdateVisibleControls();
-   float GetLFOValue(int samplesIn = 0, float forcePhase = -1);
-   float GetTargetMin() const;
-   float GetTargetMax() const;
+   double GetLFOValue(int samplesIn = 0, double forcePhase = -1);
+   double GetTargetMin() const;
+   double GetTargetMax() const;
 
    LFOSettings mLFOSettings;
