+++ Source/GridModule.h
@@ -70,7 +70,7 @@ public:
 
    //IGridControllerListener
    void OnControllerPageSelected() override;
-   void OnGridButton(int x, int y, float velocity, IGridController* grid) override;
+   void OnGridButton(int x, int y, double velocity, IGridController* grid) override;
 
    //INoteReceiver
    void PlayNote(NoteMessage note) override;
